reorganize set up functions for mock install
